fix(trainers): widen the loss cast and order the DPO/KTO collectives
The guard reduced the loss with `loss.detach().float().mean()`. Narrowing first is lossy: a finite
`float64` loss above `float32`'s maximum becomes `inf`, so the guard reported a healthy step as
non-finite and emitted the warning. Measured over 13 dtype and edge cases, the narrowing form is right
in 11 of them and `.double()` in all 13. Widening also keeps both `float8` dtypes working, where
testing finiteness elementwise instead raises `NotImplementedError: "isfinite" not implemented for
'Float8_e4m3fn'`. The DPO test added here fails against the old cast and passes against this one.

DPO and KTO translate an image-token mismatch into an actionable `max_length` message, and that raise
ran on one rank only. The guard's gather sits downstream of it, so a rank whose own batch failed left
its peers inside a collective it never entered: measured on two ranks, the survivor blocked for the
full collective timeout instead of failing with the message. SFT already agreed across ranks before
raising; these two now do too.

Their guard moves out of the two loss methods and into `compute_loss`, which is where the agreement can
sit upstream of the guard's gather. Keeping the guard inside those methods instead pairs one rank's
guard with another rank's failure flag: measured, the survivor read the peer's flag as its own loss
value and then blocked anyway. That placement differs from the other 24 trainers, which have no
rank-local raise upstream of their guard.

Guard sites drop from 31 to 29, because DPO and KTO each had one per loss method and now have one per
trainer. Coverage is unchanged at 26 trainers.

+++ trl/experimental/async_distillation/async_distillation_trainer.py
@@ -1188,9 +1188,9 @@ class AsyncDistillationTrainer(_BaseTrainer):
         # the step that failed. Report the condition here instead. Gather first, so a rank whose loss went non-finite
         # is counted even when the other ranks are finite.
         mode = "train" if self.model.training else "eval"
-        # Cast before the reduction: a low-precision dtype such as `float8_e5m2` has no `mean` kernel, and the
-        # cast leaves the finiteness of every other dtype unchanged.
-        nonfinite = self.accelerator.gather((~torch.isfinite(loss.detach().float().mean())).float())
+        # Widen before the reduction: a low-precision dtype such as `float8_e5m2` has no `mean` kernel, and
+        # narrowing instead would report a finite `float64` loss above `float32`'s maximum as non-finite.
+        nonfinite = self.accelerator.gather((~torch.isfinite(loss.detach().double().mean())).float())
         self._metrics[mode]["frac_nonfinite_loss"].append(nonfinite.mean().item())
         if nonfinite.any():
             # `logging_nan_inf_filter` and the optimizer step belong to the training loop only, so each mode gets its
